body , heading , label pages removed
